% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electricity.R
\name{electricity}
\alias{electricity}
\title{electricity}
\usage{
electricity(
  datalist,
  mapper = list(hierarchy = "main", variablemap = data.frame(label =
    c("Does this household use anything for lighting?",
    "What source of electricity is used most of the time in this household?"), variable =
    c("LIGHT01", "LIGHT03"), mappattern = c("LIGHT01", "LIGHT03")), modalitymap =
    data.frame(variable = c("LIGHT01", "LIGHT03", "LIGHT03", "LIGHT03"), label = c("yes",
    "No electricity in household", "Other, specify", "Don't know"), standard = c("1",
    "1", "96", "98"), map = c("1", "1", "96", "98")))
)
}
\arguments{
\item{datalist}{A list with all hierarchical data frame for a survey data set.
format is expected to match the Excel export synchronized from kobo to RILD
and loaded with  kobocruncher::kobo_data()}

\item{mapper}{a list providing the mapping of the variables used for the
calculation - this mapper is potentially to be adjusted in relation with deviation
between the the standard XlsForm and the contextualized dataset}
}
\value{
new calculated variable - appended or not...
}
\description{
Households lighting provides a sense of safety and security within and outside
the households after sunset.

If households lack access to electricity, especially for lighting and
connectivity, this affects the occupants' security and limits their
opportunities for socialization, learning and self-reliance.

\strong{Numerator}: Population with access to electricity, especially for lighting
and connectivity

\strong{Denominator}: Total population

\strong{Formula}: \emph{LIGHT01} = 1 & \emph{LIGHT03} != 1, 96, 98
}
\examples{
datalist <- kobocruncher::kobo_data( here::here("data-raw", "data2.xlsx"))

datalist2 <- electricity( datalist =datalist, 
         mapper = list(
            hierarchy = "main",
            variablemap = data.frame(
              label = c("Does this household use anything for lighting?",
       "What source of electricity is used most of the time in this household?"),
              variable = c("LIGHT01", 
                           "LIGHT03"),
              mappattern = c("LIGHT01", 
                           "LIGHT03") ),
            modalitymap = data.frame(
              variable = c( "LIGHT01", 
                            "LIGHT03", "LIGHT03", "LIGHT03"),
              label = c( "yes",
                        "No electricity in household", "Other, specify", "Don't know"),
              standard = c( "1",
                           "1", "96", "98"),
              map = c("yes",
                      "1", "96", "98")))  )
}
