% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_build_map.R
\name{fct_build_map}
\alias{fct_build_map}
\title{fct_build_map}
\usage{
fct_build_map(mappingfile, IndicatorRequirementFile, thisMeasureLevel)
}
\arguments{
\item{mappingfile}{full path to the xlsx file with the revise variable mapping generated with \code{fct_var_mapping()}}

\item{IndicatorRequirementFile}{path to the file where the standard mapping is depending on form version}

\item{thisMeasureLevel}{can be \code{ind} for individual level or \code{main} for household level}
}
\value{
a list
}
\description{
Quick helper to reformat the data from the mapping file into the mapping file
to use  with each indicator calculation. T
}
\details{
his needs to be done in two steps:
one for \code{ind} for individual level and
one for \code{main} for household level
}
\examples{
mappingfile <- system.file("RMS_CAPI_v2_mapping.xlsx", 
                           package = "IndicatorCalc")

IndicatorRequirementFile <- system.file("RMS_CAPI_v2_mapper.xlsx", 
                           package = "IndicatorCalc")

mappermain <- fct_build_map(mappingfile = mappingfile, 
                            IndicatorRequirementFile = IndicatorRequirementFile,
                            thisMeasureLevel = "main")

mapperind <- fct_build_map(mappingfile = mappingfile,
                            IndicatorRequirementFile = IndicatorRequirementFile,
                           thisMeasureLevel = "ind")

}
