% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_all_variable_names.R
\name{fct_get_all_variable_names}
\alias{fct_get_all_variable_names}
\title{fct_get_all_variable_names}
\usage{
fct_get_all_variable_names(datalist)
}
\arguments{
\item{datalist}{A list with all hierarchical data frame for a survey data set.
format is expected to match the Excel export synchronized from kobo to RILD
and loaded with  kobocruncher::kobo_data()}
}
\value{
dataframe with variables and list member names
}
\description{
Generate a map of all variables within the datalist obtained from RIDL -
standard Kobo extract in Excel
}
\examples{
data <- kobocruncher::kobo_data( system.file("test.xlsx",
                                             package = "IndicatorCalc"))
varname <-   fct_get_all_variable_names(datalist = data)
head(varname, 10)
}
