% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot_rbm_sdg.R
\name{fct_plot_rbm_sdg}
\alias{fct_plot_rbm_sdg}
\title{Plot SDG Indicator for a country}
\usage{
fct_plot_rbm_sdg(country = "BRA", rbm = "outcome16_2", years = c(2000, 2022))
}
\arguments{
\item{country}{iso3 code for the country (easier to recall than the M49 used in the API)}

\item{rbm}{the RBM variable name - that can match SDG}

\item{years}{years to filter the chart - for instance c(2000,2022)}
}
\value{
a ggplot2 object
}
\description{
The function is designed to help the interpretation of survey
results by providing a comparison guidelines. It displays a chart to have a
base of comparison when analyzing the results of UNHCR of UNHCR Result monitoring
survey. It pulls the values published within the UN Registry of SDG Indicator values
The function includes a mapping table between RBM/RMS and SDG and
extract the data from https://unstats.un.org/SDGAPI/swagger/ using the package SDGsR
https://drmattg.github.io/SDGsR/articles/Introduction_to_SDGsR.html
devtools::install_github("DrMattG/SDGsR", dependencies = TRUE)
}
\examples{
# fct_plot_rbm_sdg( country = "BRA", 
#           rbm = "impact2_2",
#           years = c(2000, 2022)) +
#         unhcrthemes::theme_unhcr(font_size = 10) 
}
