% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_re_map.R
\name{fct_re_map}
\alias{fct_re_map}
\title{fct_re_map}
\usage{
fct_re_map(datalist, mapper)
}
\arguments{
\item{datalist}{A list with all hierarchical data frame for a survey data set.
format is expected to match the Excel export synchronized from kobo to RILD
and loaded with  kobocruncher::kobo_data()}

\item{mapper}{a list providing the mapping of the variables used for the
calculation - this mapper is potentially to be adjusted in relation with deviation
between the the standard XlsForm and the contextualized dataset}
}
\value{
datalist
}
\description{
Take a list with a hierarchical survey as an entry - use another list to remap
variables and modalities for specific questions
}
\examples{
mapper = list(
            hierarchy = "main",
            variablemap = data.frame(
              label = c("Does this household use anything for lighting?",
                        "What source of electricity is used most of the time in this household?"),
              variable = c("LIGHT01", 
                           "LIGHT03"),
              mappattern = c("LIGHT01", 
                           "LIGHT03") ),
            modalitymap = data.frame(
              variable = c( "LIGHT01", 
                            "LIGHT03", "LIGHT03", "LIGHT03"),
              label = c( "yes",
                "No electricity in household", "Other, specify", "Don't know"),
              standard = c( "1",
                           "1", "96", "98"),
              map = c("yes",
                      "Noelec", "Other", "Dontknow")
            )
          )


## One variable is not correctly 
datalist <- list(mainhousehold = data.frame(
                group.LIGHT01 = c("yes",  "yes",  "no", "yes",  "yes",
                                  "no", "yes",  "yes",  "yes"),
                group.LIGHT03 = c("Noelec", "Other", "Dontknow", "Noelec", "Other",
                                  "Dontknow", "Nuclear", "Other", "Dontknow"))
             )
datalist <- fct_re_map(datalist = datalist, mapper = mapper )

#fct_check_map(datalist = datalist, mapper = mapper )

## Now testing on a full remap...
## Dummy data created with fct_kobo_dummy
datalist <- kobocruncher::kobo_data( system.file("dummy_RMS_CAPI_v2.xlsx",
                                                 package = "IndicatorCalc"))

## Mapping file created from Dummy data  
mappingfile <- system.file("RMS_CAPI_v2_mapping.xlsx", 
                           package = "IndicatorCalc")

IndicatorRequirementFile <- system.file("RMS_CAPI_v2_mapper.xlsx", 
                           package = "IndicatorCalc")

## and now we remap both required variables for main and ind
mappermain <- fct_build_map(mappingfile = mappingfile, 
                            IndicatorRequirementFile = IndicatorRequirementFile,
                            thisMeasureLevel = "main")
datalist <- fct_re_map(datalist = datalist, mapper = mappermain )

mapperind <- fct_build_map(mappingfile = mappingfile, 
                            IndicatorRequirementFile = IndicatorRequirementFile,
                           thisMeasureLevel = "ind")
datalist <- fct_re_map(datalist = datalist, mapper = mapperind )



## Writing this in the installation folder of the packages to run all examples
# openxlsx::write.xlsx( list ( main = as.data.frame(datalist[["main"]]),
#                              ind = as.data.frame(datalist[["ind"]]) ), 
#                       here::here("inst","dummy_RMS_CAPI_v2_mapped.xlsx"))

}
