% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_2_2.R
\name{impact_2_2}
\alias{impact_2_2}
\title{impact_2_2}
\usage{
impact_2_2(datalist)
}
\arguments{
\item{datalist}{A list with all hierarchical data frame for a survey data set.
format is expected to match the Excel export synchronized from kobo to RILD
and loaded with  kobocruncher::kobo_data()}
}
\value{
new calculated variable - appended or not...
}
\description{
\strong{Proportion of Persons of Concern residing in physically safe and secure
settlements with access to basic facilities}
}
\details{
Once electricity, healthcare, drinking water and adequate shelter are
calculated, check the values for each variable before,
calculating 2.2 impact indicator.
}
\section{inter_electricity}{
Households lighting provides a sense of safety and security within and outside
the households after sunset.

If households lack access to electricity, especially for lighting and
connectivity, this affects the occupants' security and limits their
opportunities for socialization, learning and self-reliance.

\strong{Numerator}: Population with access to electricity, especially for lighting
and connectivity

\strong{Denominator}: Total population

\strong{Formula}: \emph{LIGHT01} = 1 & \emph{LIGHT03} != 1, 96, 98
}

\section{inter_healthcare}{
Access to healthcare depends on availability of healthcare, including physical
reach, acceptability and affordability for all.
For this indicator, the focus is on the availability of healthcare system.
According to \href{https://spherestandards.org/wp-content/uploads/Sphere-Handbook-2018-EN.pdf}{The Sphere Handbook},
primary healthcare facility should be accessible within one hour's walk from dwellings.
\strong{Numerator}: Population that can reach a primary healthcare facility within one hour from dwellings
\strong{Denominator}: Total population
\strong{Formula}: \emph{HEA01} != 96, 98 & \emph{HEA03} <= 60
(reachable within one hour/60 minutes)
}

\section{inter_drinkingwater}{
Access to clean drinking water is essential for a person's survival and well
being and a precursor for achieving protection outcomes related to health,
education and economic developed.
The calculation for access drinking water is linked to SGD Indicator
\href{https://unstats.un.org/sdgs/metadata/files/Metadata-06-01-01.pdf}{6.1.1}.
The questionnaire module and the analysis guidance is taken from \href{https://mics.unicef.org/tools#analysis}{UNICEF MICS6}.

\strong{Numerator}: Population using improved sources of drinking water either in their dwelling/yard/plot or within 30 minutes round trip collection time

\strong{Denominator}: Total population

\strong{Formula}:
\emph{DWA03} < 30 (under 30 minutes), &
\emph{DWA01} != 7,9,13,96,98 &
\emph{DWA02} != 3

This basic service is calculated from the main dataset
There are three conditions as below  improved source,
in dwelling/yard/plot and reachable under 30 minutes
}

\section{inter_shelter}{
The right to access adequate housing is protected by international law.
The concept of "adequacy" means that housing is more than four walls and a
roof as indicated in \href{https://spherestandards.org/wp-content/uploads/Sphere-Handbook-2018-EN.pdf}{The Sphere Handbook}.
Habitable housing primarily refers to the fact that the housing should provide
protection from cold, damp, heat, rain, wind, and other threats to health,
structural hazards, and disease vectors and it should not be overcrowded.
As shelter/housing is primarily a contextual element, there may be
discrepancies from country to country on how this data is measured.

Adequate shelter is measured based on having improved material for the
dwelling as indicated in \href{https://dhsprogram.com/pubs/pdf/AS61/AS61.pdf}{DHS}
publication on housing conditions which is also used by \href{https://mics.unicef.org/tools}{MICS6}.

Overcrowding is also used which occurs if there are more than three people
per habitable room as defined by \href{https://www.ncbi.nlm.nih.gov/books/NBK535289/table/ch3.tab2/}{UN-Habitat}.

\if{html}{\out{<div class="sourceCode">}}\preformatted{**Numerator**: Population that have access to adequate housing
**Denominator**: Total population
}\if{html}{\out{</div>}}

\strong{Formula}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{ *DWE01* = 1,2 &
 *DWE02* = 3,4,5,6,7,8,9 & 
 *DWE03* = 8,9,10,11,12,13 & 
 *DWE04* = 10,11,12,13,14,15 & 
 crowding (*HH01*/*DWE05*) \\<= 3
}\if{html}{\out{</div>}}

Adequate shelter is calculated from the main dataset
classify as habitable when improved/adequate shelter
}

\section{Compile alll}{
Once all variables are correctly calculated, we can compute the final
variable for impact 2.2 indicator.
\strong{Numerator}: Population residing in physically safe and secure settlements with access to basic facilities
\strong{Denominator}: Total population
\strong{Formula}: \emph{shelter}=1 & \emph{electricity}=1 & \emph{healthcare}=1 & \emph{drinkingwater}=1

Impact 2.2 is "1" if all services above are accessible
}

\examples{
## data, cf example  fct_re_map()
datalist <- kobocruncher::kobo_data( system.file("dummy_RMS_CAPI_v2_mapped.xlsx", 
                                                 package = "IndicatorCalc"))
## Apply calculation
datalist <-  impact_2_2(datalist)

table(datalist[["main"]]$impact2_2, useNA = "ifany")

fct_plot_indic_donut(indicator = datalist[["main"]]$impact2_2,
                     iconunicode = "f140") 


## Can get the details as well

table(datalist[["main"]]$electricity, useNA = "ifany")
fct_plot_indic_donut(indicator = datalist[["main"]]$electricity,
                     iconunicode = "f0e7") 

table(datalist[["main"]]$healthcare, useNA = "ifany")
fct_plot_indic_donut(indicator = datalist[["main"]]$healthcare,
                     iconunicode = "f479") 

 
table(datalist[["main"]]$drinkingwater, useNA = "ifany")
fct_plot_indic_donut(indicator = datalist[["main"]]$drinkingwater,
                     iconunicode = "e006") 

## Check intermediary variables
table(datalist[["main"]]$dwa_cond1, useNA = "ifany")
table(datalist[["main"]]$reachableU30, useNA = "ifany")
table(datalist[["main"]]$DWA02, useNA = "ifany")
table(datalist[["main"]]$dwa_cond2, useNA = "ifany")

# Tabulate
table(datalist[["main"]]$dwe01_cat, useNA = "ifany")
table(datalist[["main"]]$dwe02_cat, useNA = "ifany")
table(datalist[["main"]]$dwe03_cat, useNA = "ifany")
table(datalist[["main"]]$dwe04_cat, useNA = "ifany")
table(datalist[["main"]]$dwe05_cat, useNA = "ifany")
table(datalist[["main"]]$shelter, useNA = "ifany")
#plot
fct_plot_indic_donut(datalist[["main"]]$shelter,
                     iconunicode = "e54f") 

}
