% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inter_drinkingwater.R
\name{inter_drinkingwater}
\alias{inter_drinkingwater}
\title{inter_drinkingwater}
\usage{
inter_drinkingwater(
  datalist,
  mapper = list(hierarchy = "main", variablemap = data.frame(label =
    c("What is the main source of drinking water for this household?",
    "Where is this source located?", "Unit used to measure time to access",
    "How long does it take to go there, get water, and come back,\\n                  including waiting time?"),
    variable = c("DWA01", "DWA02", "DWA03a", "DWA03b"), mappattern = c("DWA01", "DWA02",
    "DWA03a", "DWA03b")), modalitymap = data.frame(variable = c("DWA01", "DWA01",
    "DWA01", "DWA01", "DWA01", "DWA02", 
     "DWA02", "DWA02", "DWA03a", "DWA03a"),
    label = c("Unprotected Dug Well", "Unprotected Spring",
    "Surface Water (River, Stream, Pond, Dam, Canal)", "Other, specify", "Don't know",
    "In Own Dwelling", "In Own Yard/Plot", "Elsewhere", "Minutes", "Hours"), standard =
    c("7", "9", "13", "96", "98", "1", "2", "3", "1", "2"), map = c("7", "9", "13", "96",
    "98", "1", "2", "3", "1", "2")))
)
}
\arguments{
\item{datalist}{A list with all hierarchical data frame for a survey data set.
format is expected to match the Excel export synchronized from kobo to RILD
and loaded with  kobocruncher::kobo_data()}

\item{mapper}{a list providing the mapping of the variables used for the
calculation - this mapper is potentially to be adjusted in relation with deviation
between the the standard XlsForm and the contextualized dataset}
}
\value{
new calculated variable - appended or not...
}
\description{
Access to clean drinking water is essential for a person's survival and well
being and a precursor for achieving protection outcomes related to health,
education and economic developed.
The calculation for access drinking water is linked to SGD Indicator
\href{https://unstats.un.org/sdgs/metadata/files/Metadata-06-01-01.pdf}{6.1.1}.
The questionnaire module and the analysis guidance is taken from \href{https://mics.unicef.org/tools#analysis}{UNICEF MICS6}.

\strong{Numerator}: Population using improved sources of drinking water either in their dwelling/yard/plot or within 30 minutes round trip collection time

\strong{Denominator}: Total population

\strong{Formula}:
\emph{DWA03} < 30 (under 30 minutes), &
\emph{DWA01} != 7,9,13,96,98 &
\emph{DWA02} != 3

This basic service is calculated from the main dataset
There are three conditions as below  improved source,
in dwelling/yard/plot and reachable under 30 minutes
}
\examples{
datalist <- kobocruncher::kobo_data( system.file("test.xlsx",
                                                 package = "IndicatorCalc"))

## in the contextualised form - DWA03a has been skipped and all results are in min... 
## only manual transformation can adjust this before we use the mapper..

datalist[["main"]]$DWA03a  <- "1" 

datalist[["main"]]$DWA03b <- 
  datalist[["main"]]$VulnerabilityScoring.BasicNeeds.DWA03

# now the mapper
mapper <-  list(
            hierarchy = "main",
            variablemap = data.frame(
              label = c(
   "What is the main source of drinking water for this household?",
  "Where is this source located?",
  "Unit used to measure time to access",
  "How long does it take to go there, wait get water, and come back?"),
              variable = c("DWA01", 
                           "DWA02", 
                           "DWA03a", 
                           "DWA03b" ),
              mappattern = c("DWA01", 
                           "DWA02", 
                           "DWA03a", 
                           "DWA03b" ) ),
            modalitymap = data.frame(
             variable = c("DWA01",  "DWA01", "DWA01","DWA01",  "DWA01", 
                   "DWA02","DWA02", "DWA02",
                   "DWA03a","DWA03a"),
             label = c( 
                ##DWA01
                "Unprotected Dug Well", 
                "Unprotected Spring",
                "Surface Water (River, Stream, Pond, Dam, Canal)",
                "Other, specify",
                "Don't know",
                ##DWA02
                "In Own Dwelling", 
                "In Own Yard/Plot",
                "Elsewhere",
                ## DWA03a
                "Minutes", 
                "Hours"    ),
             standard = c( "7", "9", "13", "96", "98",
                           "1", "2", "3",
                           "1", "2"),
             map = c( "7", "9", "13", "96", "98",
                       "1", "2", "3",
                        "1", "2") ) 
         )
  
datalist <- inter_drinkingwater(datalist, mapper )

## Indicator summary
table(datalist[["main"]]$drinkingwater, useNA = "ifany")
fct_plot_indic_donut(indicator = datalist[["main"]]$drinkingwater,
                     iconunicode = "e006") 

## Check auxilliary
table(datalist[["main"]]$dwa_cond1, useNA = "ifany")
table(datalist[["main"]]$reachableU30, useNA = "ifany")
table(datalist[["main"]]$DWA02, useNA = "ifany")
table(datalist[["main"]]$dwa_cond2, useNA = "ifany")
}
