% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inter_healthcare.R
\name{inter_healthcare}
\alias{inter_healthcare}
\title{inter_healthcare}
\usage{
inter_healthcare(
  datalist,
  mapper = list(hierarchy = "main", variablemap = data.frame(label =
    c("In general, when anyone in your household is sick,  where do they go to seek care?",
    "How long does it take to go there when you use the mode of transport that you mentioned above?"),
    variable = c("HEA01", "HEA03"), mappattern = c("HEA01", "HEA03")), modalitymap =
    data.frame(variable = c("HEA01", "HEA01"), label = c("Other, specify", "Don't know"),
    standard = c("96", "98"), map = c("96", "98")))
)
}
\arguments{
\item{datalist}{A list with all hierarchical data frame for a survey data set.
format is expected to match the Excel export synchronized from kobo to RILD
and loaded with  kobocruncher::kobo_data()}

\item{mapper}{a list providing the mapping of the variables used for the
calculation - this mapper is potentially to be adjusted in relation with deviation
between the the standard XlsForm and the contextualized dataset}
}
\value{
new calculated variable - appended or not...
}
\description{
Access to healthcare depends on availability of healthcare, including physical
reach, acceptability and affordability for all.
For this indicator, the focus is on the availability of healthcare system.
According to \href{https://spherestandards.org/wp-content/uploads/Sphere-Handbook-2018-EN.pdf}{The Sphere Handbook},
primary healthcare facility should be accessible within one hour's walk from dwellings.
\strong{Numerator}: Population that can reach a primary healthcare facility within one hour from dwellings
\strong{Denominator}: Total population
\strong{Formula}: \emph{HEA01} != 96, 98 & \emph{HEA03} <= 60
(reachable within one hour/60 minutes)
}
\examples{
datalist <- kobocruncher::kobo_data( system.file("test.xlsx",
                                                 package = "IndicatorCalc"))

mapper <-  list(
            hierarchy = "main",
            variablemap = data.frame(
              label = c(
"In general, when anyone in your household is sick, where do they go to seek care?",
"How long it takes to go there when you use the mode of transport mentioned above?"),
              variable = c("HEA01", 
                           "HEA03"),
              mappattern = c("HEA01", 
                           "HEA03") ),
            modalitymap = data.frame(
              variable = c( "HEA01", "HEA01" ),
              label = c(  "Other, specify", "Don't know"),
              standard = c("96", "98" ),
              map = c("96", "98" )))

datalist <- inter_healthcare(datalist, mapper )

table(datalist[["main"]]$healthcare)
fct_plot_indic_donut(indicator = datalist[["main"]]$healthcare,
                     iconunicode = "f479") 

}
