% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inter_shelter.R
\name{inter_shelter}
\alias{inter_shelter}
\title{inter_shelter}
\usage{
inter_shelter(
  datalist,
  mapper = list(hierarchy = "main", variablemap = data.frame(label =
    c("What type of dwelling does the household live in?",
    "Main material of the dwelling floor", "Main material of the roof",
    "Main material of the exterior walls",
    "How many separate rooms do the members of your household occupy?",
    "What is the total number of persons in this household?"), variable = c("DWE01",
    "DWE02", "DWE03", "DWE04", "DWE05", "HH01"), mappattern = c("DWE01", "DWE02",
    "DWE03", "DWE04", "DWE05", "HH01")), modalitymap = data.frame(variable = c("DWE01", 

        "DWE01", "DWE02", "DWE02", "DWE02", "DWE03", "DWE03", "DWE03", "DWE03", "DWE03",
    "DWE03", "DWE04", "DWE04", "DWE04", "DWE04", "DWE04", "DWE04"), label =
    c("Apartment", "House", "Earth/sand", "Dung", "Other (Specify)", "Metal/tin", "Wood",
    "Calamine/Cement fibre", "Ceramic tiles", "Cement", "Roofing shingles", "Cement",
    "Stone with lime/ cement", "Bricks", "Cement blocks", "Covered adobe",
    "Wood planks/shingles"), standard = c("1", "2", "1", "2", "96", "8", "9", "10", "11",
    "12", "13", "10", "11", 
     "12", "13", "14", "15"), map = c("1", "2", "1", "2",
    "96", "8", "9", "10", "11", "12", "13", "10", "11", "12", "13", "14", "15")))
)
}
\arguments{
\item{datalist}{A list with all hierarchical data frame for a survey data set.
format is expected to match the Excel export synchronized from kobo to RILD
and loaded with  kobocruncher::kobo_data()}

\item{mapper}{a list providing the mapping of the variables used for the
calculation - this mapper is potentially to be adjusted in relation with deviation
between the the standard XlsForm and the contextualized dataset}
}
\value{
new calculated variable - appended or not...
}
\description{
The right to access adequate housing is protected by international law.
The concept of "adequacy" means that housing is more than four walls and a
roof as indicated in \href{https://spherestandards.org/wp-content/uploads/Sphere-Handbook-2018-EN.pdf}{The Sphere Handbook}.
Habitable housing primarily refers to the fact that the housing should provide
protection from cold, damp, heat, rain, wind, and other threats to health,
structural hazards, and disease vectors and it should not be overcrowded.
As shelter/housing is primarily a contextual element, there may be
discrepancies from country to country on how this data is measured.

Adequate shelter is measured based on having improved material for the
dwelling as indicated in \href{https://dhsprogram.com/pubs/pdf/AS61/AS61.pdf}{DHS}
publication on housing conditions which is also used by \href{https://mics.unicef.org/tools}{MICS6}.

Overcrowding is also used which occurs if there are more than three people
per habitable room as defined by \href{https://www.ncbi.nlm.nih.gov/books/NBK535289/table/ch3.tab2/}{UN-Habitat}.

\if{html}{\out{<div class="sourceCode">}}\preformatted{**Numerator**: Population that have access to adequate housing
**Denominator**: Total population
}\if{html}{\out{</div>}}

\strong{Formula}:
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ *DWE01* = 1,2 &
 *DWE02* = 3,4,5,6,7,8,9 & 
 *DWE03* = 8,9,10,11,12,13 & 
 *DWE04* = 10,11,12,13,14,15 & 
 crowding (*HH01*/*DWE05*) \\<= 3
}\if{html}{\out{</div>}}

Adequate shelter is calculated from the main dataset
classify as habitable when improved/adequate shelter
}
\examples{
datalist <- kobocruncher::kobo_data( system.file("test.xlsx", 
                                                 package = "IndicatorCalc"))

mapper <- list(
            hierarchy = "main",
            variablemap = data.frame(
              label = c(
                "What type of dwelling does the household live in?",
"Main material of the dwelling floor",
"Main material of the roof",
"Main material of the exterior walls",
"How many separate rooms do the members of your household occupy?",
"What is the total number of persons in this household?"),
              variable = c("DWE01","DWE02","DWE03","DWE04","DWE05", 
                           "HH01"),
              mappattern = c("DWE01","DWE02","DWE03","DWE04","DWE05", 
                           "progres_groupsize") ),
            modalitymap = data.frame(
              variable = c( "DWE01","DWE01",
                            "DWE02","DWE02","DWE02",
                            "DWE03","DWE03","DWE03","DWE03","DWE03","DWE03",
                            "DWE04","DWE04","DWE04","DWE04","DWE04","DWE04"),
              label = c(  "Apartment", "House", # DWE01
                         "Earth/sand", "Dung", "Other (Specify)", #DWE02
                         
                         "Metal/tin", "Wood", "Calamine/Cement fibre", 
                         "Ceramic tiles", "Cement", "Roofing shingles",#DWE03
                         
                         "Cement", "Stone with lime/ cement", "Bricks", 
                         "Cement blocks", "Covered adobe", "Wood planks/shingles" # DWE04
                         ),
              standard = c( "1","2",
                           "1", "2","96",
                           "8","9","10","11","12","13",
                           "10","11","12","13","14","15"),
              map = c("1","2",
                           "1", "2","96",
                           "8","9","10","11","12","13",
                           "10","11","12","13","14","15"))) 
## Calculate
datalist <-  inter_shelter(datalist, mapper)
# Tabulate
table(datalist[["main"]]$dwe01_cat)
table(datalist[["main"]]$dwe02_cat)
table(datalist[["main"]]$dwe03_cat)
table(datalist[["main"]]$dwe04_cat)
table(datalist[["main"]]$dwe05_cat)
table(datalist[["main"]]$shelter)
#plot
fct_plot_indic_donut(datalist[["main"]]$shelter,
                     iconunicode = "e54f") 
}
