% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_dummy.R
\name{kobo_dummy}
\alias{kobo_dummy}
\title{kobo_dummy}
\usage{
kobo_dummy(form, n = 384, file)
}
\arguments{
\item{form}{file path of the XlsForm}

\item{n}{number of main records to be generated}

\item{file}{file as xlsx where to save to the}

\item{nrepeat}{max random number of repeat records to be generated
when repeat_count is not mentionned}
}
\value{
a data list with a series of dummy data
}
\description{
Automatically produce an dummy dataset in line with the structure
of an XlsForm
Making decisions about research design and analysis strategies is often difficult before data is collected,
because it is hard to imagine the exact form data will take.
This function helps imagine what data will look like before they collect it.
samplesize is set per defautl at 500 records

Supported Features:
\itemize{
\item Generate a data set
\item respects ODK structure "\code{relevant}" skip logic
(Some advanced functionality such as "coalesce()" not covered)
\item respects "\code{constraint}"
\item respects "\code{repeat}" adds ID column to link hierarchical data based on "\code{repeat_count}"
This function is a rewriting of
https://unhcr.github.io/koboloadeR/docs/reference/kobo_dummy.html
it also  build on  https://thinkr-open.github.io/fakir/ and
https://docs.ropensci.org/charlatan/
}
}
\examples{
form <- system.file("RMSCAPI.xlsx", package = "IndicatorCalc")
datalist <- kobo_dummy(form,
                       n = 384,
                       file = NULL)

## Save this to use it for testing the package...

openxlsx::write.xlsx(datalist, here::here("inst", "demo_data.xlsx"))
}
