% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome10_1.R
\name{outcome10_1}
\alias{outcome10_1}
\title{outcome10_1}
\usage{
outcome10_1(datalist)
}
\arguments{
\item{datalist}{A list with all hierarchical data frame for a survey data set.
format is expected to match the Excel export synchronized from kobo to RILD
and loaded with  kobocruncher::kobo_data()}
}
\value{
new calculated variable - appended or not...
}
\description{
\strong{Proportion of children aged 9 months to five years who have received measles vaccination}
}
\details{
Coverage of measles vaccination in children in the target age group of 9 months to 5 years of age.
Measles vaccination is an essential preventive primary care intervention to protect children from measles infection.

Standard module is from \href{https://mics.unicef.org/tools#survey-design}{UNICEF MICS6} Children under 5 module.
The calculation of this module is also aligned with UNICEF MICS6.\tabular{c}{
   Standard Questions \cr
   MMR01-MMR04 \cr
}


\strong{Numerator}: Total number of children 9 months to 5 years who have received a measles containing vaccine (measles or MMR- Measles Mumps and Rubella)

\strong{Denominator}: Total number of children 9 months to 5 years

\strong{Formula}: *MMR03=*1 / Total number of children 9 months to 5 years

This comes from the main dataset

MICS TC.1.1 UNICEF calculates on the first dose received##
Children who had et least one dose of measles vaccination
}
\examples{
## data, cf example  fct_re_map()
datalist <- kobocruncher::kobo_data( system.file("dummy_RMS_CAPI_v2_mapped.xlsx", 
                                                 package = "IndicatorCalc"))

## Apply indicator function on datalist
datalist <- outcome10_1(datalist)

## Visualise value
fct_plot_indic_donut(indicator = datalist[["ind"]]$outcome10_1,
                     iconunicode = "f140")   
}
