% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome12_1.R
\name{outcome12_1}
\alias{outcome12_1}
\title{outcome12_1}
\usage{
outcome12_1(datalist)
}
\arguments{
\item{datalist}{A list with all hierarchical data frame for a survey data set.
format is expected to match the Excel export synchronized from kobo to RILD
and loaded with  kobocruncher::kobo_data()}
}
\value{
new calculated variable - appended or not...
}
\description{
\strong{Proportion of Persons of Concern using at least basic drinking water services}
}
\details{
This indicator is defined as the percentage of PoCs using at least basic water services.
Access to clean drinking water is essential for a person's survival and well being and a precursor for achieving protection outcomes related to health, education and economic developed.
The calculation for access drinking water is linked to SGD Indicator \href{https://unstats.un.org/sdgs/metadata/files/Metadata-06-01-01.pdf}{6.1.1}.
The questionnaire module and the analysis guidance is taken from \href{https://mics.unicef.org/tools#analysis}{UNICEF MICS6}.\tabular{c}{
   Standard Questions \cr
   DWA01-DWA04 \cr
}


\strong{Numerator}: Population using improved sources of drinking water either in their dwelling/yard/plot or within 30 minutes round trip collection time

\strong{Denominator}: Total population

\strong{Formula}:
\emph{DWA03} \< 30 (under 30 minutes), &
\emph{DWA01} !=7,9,13,96,98 &
\emph{DWA02} !=3

This comes from the main dataset
There are two conditions as below improved source, in dwelling/yard/plot or reachable under 30 minutes
}
\examples{
## data, cf example  fct_re_map()
datalist <- kobocruncher::kobo_data( system.file("dummy_RMS_CAPI_v2_mapped.xlsx", 
                                                 package = "IndicatorCalc"))
## Apply indicator function on datalist
datalist <- outcome12_1(datalist  )

## Visualise value
fct_plot_indic_donut(indicator = datalist[["main"]]$outcome12_1,
                     iconunicode = "f140")   
}
