% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome12_2.R
\name{outcome12_2}
\alias{outcome12_2}
\title{outcome12_2}
\usage{
outcome12_2(datalist)
}
\arguments{
\item{datalist}{A list with all hierarchical data frame for a survey data set.
format is expected to match the Excel export synchronized from kobo to RILD
and loaded with  kobocruncher::kobo_data()}
}
\value{
new calculated variable - appended or not...
}
\description{
\strong{Proportion of Persons of Concern with access to a safe household toilet}
}
\details{
This indicator measures the proportion of persons of concerns with access to at least basic sanitation services -- toilets -- that are not shared with other households.
A toilet is defined as a basic sanitation facility.
This indicator is linked to SGD indicator \href{https://sdg-tracker.org/water-and-sanitation#:~:text=Definition\%3A\%20Indicator\%206.2.,at\%20least\%20basic\%20handwashing\%20facilities.}{6.2.1}.
The standard module is taken from \href{https://mics.unicef.org/tools#analysis}{UNICEF MICS6} main household questionnaire.
Calculation of the indicator is based on \href{https://mics.unicef.org/tools#analysis}{MICS6 analysis tools}.\tabular{c}{
   Standard Questions \cr
   TOI01-TOI05 \cr
}


\strong{Numerator}: Total population with access to sanitation facility at their household

\strong{Denominator}: Total population

\strong{Formula}: \emph{TOI01}=1,2,3,4,5,6,7,9 & (\emph{TOI02} =1 & \emph{TOI03}=1,2,3,4) & \emph{TOI05}=1
This comes from the main dataset
MICS calculation WS3.1/WS3.4
}
\examples{
## data, cf example  fct_re_map()
datalist <- kobocruncher::kobo_data( system.file("dummy_RMS_CAPI_v2_mapped.xlsx", 
                                                 package = "IndicatorCalc"))
## Apply indicator function on datalist
datalist <- outcome12_2(datalist)

## Visualise value
fct_plot_indic_donut(indicator = datalist[["main"]]$outcome12_2,
                     iconunicode = "f140")   
}
