% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome13_1.R
\name{outcome13_1}
\alias{outcome13_1}
\title{outcome13_1}
\usage{
outcome13_1(datalist)
}
\arguments{
\item{datalist}{A list with all hierarchical data frame for a survey data set.
format is expected to match the Excel export synchronized from kobo to RILD
and loaded with  kobocruncher::kobo_data()}
}
\value{
new calculated variable - appended or not...
}
\description{
\strong{Proportion of Persons of Concern with an account at a bank or other financial institution or with a mobile-money-service provider}
}
\details{
This indicator measures the percentage of persons of concern (ages 15+) who report having an account (by themselves or together with someone else) at a bank or another type of financial institution or personally using a mobile money service in the past 12 months in the country of asylum or habitual residence (for returnees, countries of origin are included).

The methodology is taken from \href{https://www.worldbank.org/en/publication/globalfindex}{The Global Findex Database} which is developed by the World Bank.
This indicator is also linked to SGD Indicator \href{https://unstats.un.org/sdgs/metadata/files/Metadata-08-10-02.pdf}{8.10.2}.\tabular{c}{
   Standard Questions \cr
   BANK01-BANK05 \cr
}


\strong{Numerator}: Total population having a personal mobile or bank account

\strong{Denominator}: Total population

\strong{Formula}: \emph{BANK01}=1 \| \emph{BANK02}=1 \| \emph{BANK03}=1 \|\emph{BANK05}=1
This indicator comes from main dataset based on the respondent randomly selected for individual level
include if the respondent has an account on their own/with someone else, ATM card, personal bank card
or personally used phone to send money (proxy)
}
\examples{
## data, cf example  fct_re_map()
datalist <- kobocruncher::kobo_data( system.file("dummy_RMS_CAPI_v2_mapped.xlsx", 
                                                 package = "IndicatorCalc"))
## Apply indicator function on datalist
datalist <- outcome13_1(datalist )

## Visualise value
fct_plot_indic_donut(indicator = datalist[["main"]]$outcome13_1,
                     iconunicode = "f140")   
}
