% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome13_3.R
\name{outcome13_3}
\alias{outcome13_3}
\title{outcome13_3#'}
\usage{
outcome13_3(datalist)
}
\arguments{
\item{datalist}{A list with all hierarchical data frame for a survey data set.
format is expected to match the Excel export synchronized from kobo to RILD
and loaded with  kobocruncher::kobo_data()}
}
\value{
new calculated variable - appended or not...
}
\description{
\strong{Proportion of Persons of Concern (working age) who are unemployed}
The unemployment rate is a standard indicator to measure the efficiency and effectiveness of an economy.
For Persons of Concern, access to employment is central for self-reliance, socio-economic inclusion and dignity and often restricted through laws, policies or practical barriers.
This indicator is also linked to SGD \href{https://unstats.un.org/sdgs/metadata/files/Metadata-08-05-02.pdf}{8.5.2}.
The standard questionnaire module is from \href{https://www.unhcr.org/5ea81b954}{Standardized Employment Module} of UNHCR.
The calculations are also done based on the same guidance.\tabular{c}{
   Standard Questions \cr
   UNEM01-UNEM10 \cr
}
}
\details{
\strong{Numerator}: Unemployed working age population within a country

\strong{Denominator}: Total working age population who are employed and unemployed within a country (often referred to as labour force)

\strong{Formula}: \emph{UNEM01}=1 \| (\emph{UNEM02}=1 & \emph{UNEM07}=3) \| \emph{UNEM04}=1 \| (\emph{UNEM02}=1 & \emph{UNEM07}=1 &( \emph{UNEM08}=1,2)) \|(\emph{UNEM05}=1 & \emph{UNEM06}=3) \| ( \emph{UNEM05}=1 & ( \emph{UNEM06}=1,2 \| \emph{UNEM08}=1,2)
This indicator comes from main dataset based on the respondent randomly selected for individual level

#Numerator: Those of working age who were not in employment, looked for employment in the past 30 days and were available to take up employment
}
\examples{
## data, cf example  fct_re_map()
datalist <- kobocruncher::kobo_data( system.file("dummy_RMS_CAPI_v2_mapped.xlsx", 
                                                 package = "IndicatorCalc"))
## Apply indicator function on datalist
datalist <- outcome13_3(datalist )

table( datalist[["main"]]$outcome13_3, useNA = "ifany")
## Visualise value
fct_plot_indic_donut(indicator = datalist[["main"]]$outcome13_3,
                     iconunicode = "f140")   
}
