% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome14_1.R
\name{outcome14_1}
\alias{outcome14_1}
\title{outcome14_1}
\usage{
outcome14_1(datalist)
}
\arguments{
\item{datalist}{A list with all hierarchical data frame for a survey data set.
format is expected to match the Excel export synchronized from kobo to RILD
and loaded with  kobocruncher::kobo_data()}
}
\value{
new calculated variable - appended or not...
}
\description{
\strong{Proportion of returnees with legally recognized identity documents or credentials}
}
\details{
This indicator measures the proportion of returned refugees who possess legally recognized and valid identity documents or credentials to support their return .

Commonly in the context of return, returnees require civil documentation or credentials and inclusion in or updating of civil registries , as well as access to services.

\emph{\strong{Definitions}}

• Identity document or credential is any document or credential which may be used as proof of identity, which may also include reference to the individuals' legal status and associated rights vis-à-vis the host State and/or UNHCR.\tabular{c}{
   Standard Questions \cr
   REG01 - REG02 - REG03 / REG05 - REG06 \cr
}


\strong{Numerator}: Total number of returnees with with legally recognized identity documents or credentials

\strong{Denominator}: Total number of returnees

\strong{Formula}: \emph{REG01}=1 \| \emph{REG02}=1 \| \emph{REG03}=1 \| \emph{REG05}=1 \| \emph{REG06}=1

This indicator comes from the individual dataset
Calculate valid identity documents for under 5 with REG05 and REG06 variables
}
\examples{
## data, cf example  fct_re_map()
datalist <- kobocruncher::kobo_data( system.file("dummy_RMS_CAPI_v2_mapped.xlsx", 
                                                 package = "IndicatorCalc"))

## Apply indicator function on datalist
datalist <- outcome14_1(datalist )

## Visualise value
fct_plot_indic_donut(indicator = datalist[["ind"]]$outcome14_1,
                     iconunicode = "f140")   
}
