% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome16_1.R
\name{outcome16_1}
\alias{outcome16_1}
\title{outcome16_1}
\usage{
outcome16_1(datalist)
}
\arguments{
\item{datalist}{A list with all hierarchical data frame for a survey data set.
format is expected to match the Excel export synchronized from kobo to RILD
and loaded with  kobocruncher::kobo_data()}
}
\value{
new calculated variable - appended or not...
}
\description{
\strong{Proportion of Persons of Concern with secure tenure rights and/or property rights to housing and/or land}

This indicator measures the proportion of persons of concern that have secure tenure rights to housing and/or land.
Security of tenure means that persons of concern can live in their homes without fear of forced eviction, whether in communal settlement situations, informal settlements, host communities or after return.
}
\details{
This indicator is linked to SGD Indicator \href{https://unstats.un.org/sdgs/metadata/files/Metadata-01-04-02.pdf}{1.4.2}. The further guidance can be found \href{https://gltn.net/download/measuring-individuals-rights-to-land-an-integrated-approach-to-data-collection-for-sdg-indicators-1-4-2-and-5-a-1-english/?wpdmdl=16316&refresh=5efb342458df61593521188}{here}

\emph{Concept}
\itemize{
\item Secure tenure rights: comprised of two sub-components: (i) legally recognized documentation and (ii) perception of the security of tenure, which are both necessary to provide a full measurement of tenure security.
\item Legally recognized documentation: Legal documentation of rights refers to the recording and publication of information on the nature and location of land, rights and right holders in a form that is recognized by government, and is therefore official.
\item Perceived security of tenure: Perception of tenure security refers to an individual's perception of the likelihood of involuntary loss of land, such as disagreement of the ownership rights over land or ability to use it, regardless of the formal status and can be more optimistic or pessimistic.
}\tabular{c}{
   Standard Questions \cr
   DWE06-DWE07 & DWE10-DWE11 \cr
}


\strong{Numerator}: Total population with secure tenure rights to housing and/or land

\strong{Denominator}: Total population
\strong{Formula}: \emph{DWE011} = 1,2 & \emph{DWE10} = 1,2,3,4,5,6 & \emph{DWE06} != 9, 96,98,99 & \emph{DWE07} = 9, 96,98,99
This indicator is calculated from the main dataset
}
\examples{
## data, cf example  fct_re_map()
datalist <- kobocruncher::kobo_data( system.file("dummy_RMS_CAPI_v2_mapped.xlsx", 
                                                 package = "IndicatorCalc"))

## Apply indicator function on datalist
datalist <- outcome16_1(datalist )

## Visualise value
fct_plot_indic_donut(indicator = datalist[["main"]]$outcome16_1,
                     iconunicode = "f140")    
}
