% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome16_2.R
\name{outcome16_2}
\alias{outcome16_2}
\title{outcome16_2}
\usage{
outcome16_2(datalist)
}
\arguments{
\item{datalist}{A list with all hierarchical data frame for a survey data set.
format is expected to match the Excel export synchronized from kobo to RILD
and loaded with  kobocruncher::kobo_data()}
}
\value{
new calculated variable - appended or not...
}
\description{
\strong{Proportion of Persons of Concern covered by social protection floors/systems}
This indicator measures the proportion of persons of concern who receive government social protection benefits.
Social protection is a set of policies and programmes aimed at preventing or
protecting all people against poverty, vulnerability and social exclusion
throughout their life-course, with particular emphasis on vulnerable groups.
For persons of concern, being part of a social protection system is a measure
of integration and stability. This indicator is linked to SGD indicator
\href{https://unstats.un.org/sdgs/metadata/files/Metadata-01-03-01a.pdf}{1.3.1}.

This indicator focuses on the social benefits received from the government in which UNHCR might be supporting UNHCR.\tabular{c}{
   Standard Questions \cr
   SPF01 \cr
}


\strong{Numerator}: Total population covered by social protection floors/systems
\strong{Denominator}: Total population
\strong{Formula}: \emph{SPF01}=1

This indicator is calculated from the main dataset
If PoC has covered by at least one of the social protection floors/systems
}
\examples{
## data, cf example  fct_re_map()
datalist <- kobocruncher::kobo_data( system.file("dummy_RMS_CAPI_v2_mapped.xlsx", 
                                                 package = "IndicatorCalc"))
## Apply indicator function on datalist
datalist <- outcome16_2(datalist)

table( datalist[["main"]]$outcome16_2, useNA = "ifany")
## Visualise value
fct_plot_indic_donut(indicator = datalist[["main"]]$outcome16_2,
                     iconunicode = "f140")   
}
