% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome1_2.R
\name{outcome1_2}
\alias{outcome1_2}
\title{outcome1_2}
\usage{
outcome1_2(datalist)
}
\arguments{
\item{datalist}{A list with all hierarchical data frame for a survey data set.
format is expected to match the Excel export synchronized from kobo to RILD
and loaded with  kobocruncher::kobo_data()}
}
\value{
new calculated variable - appended or not...
}
\description{
\strong{Proportion of children under 5 years of age whose births have been registered with a civil authority}
}
\details{
This indicator measures the proportion of children under 5 years old whose births
have been registered with a civil authority and civil documentation has been issued.
This is directly linked to SGD indicator \href{https://unstats.un.org/sdgs/metadata/files/Metadata-16-09-01.pdf}{16.9.1}.

The standard module for this indicator is taken from MICS and adjusted to UNHCR context.

\strong{Definitions}

•\strong{Birth registration} refers to the registration of new births by civil
authorities and completion of the process through the issuance of a birth certificate.
This also includes documents issued by UNHCR or other relevant organizations
when \strong{given the authority by the State and these documents are recognized by national authorities}.

\strong{Clarifications}

•Birth notifications or other hospital records and records from midwives or
traditional birth attendants, \strong{issued solely by UNHCR or its partners shall
not be considered} as birth certificates although they are important sources
for establishing the total number of births.
Operations are encouraged to track both the number of birth notifications and
birth registrations but for the purpose of this indicator should report
the number of births registered.

• The standard indicator used in DHS, MICS and RMS to report on birth
registration refers to the percentage of children under age 5 (0-59 months)
with a birth certificate, regardless of whether or not it was seen by the
interviewer, or whose birth was reported as registered with civil
authorities at the time of survey.

\if{html}{\out{<div class="sourceCode">}}\preformatted{| Standard Questions |
|:------------------:|
|   REG03 - REG04    |

**Numerator**: Number of children under 5 years old who are registered with civil authorities
**Denominator**: Total number of children under 5 years old
**Formula**: (*REG03=1 \\| REG04=1) / Number of children under 5*

This indicator comes from the individual dataset
 ind$REG03 - birth certificate
 ind$REG04 - birth has been registered
}\if{html}{\out{</div>}}
}
\examples{
## data, cf example  fct_re_map()
datalist <- kobocruncher::kobo_data( system.file("dummy_RMS_CAPI_v2_mapped.xlsx", 
                                                 package = "IndicatorCalc"))

## Apply indicator function on datalist
datalist <- outcome1_2(datalist)
table(datalist[["ind"]]$outcome1_2, useNA = "ifany")
table(datalist[["ind"]]$less_than_5, useNA = "ifany")
table(datalist[["ind"]]$HH07, useNA = "ifany")
barplot(as.integer(datalist[["ind"]]$HH07))
table(datalist[["ind"]]$birthCertificate, useNA = "ifany")
table(datalist[["ind"]]$birthRegistered, useNA = "ifany")
## Visualise value
fct_plot_indic_donut(indicator = datalist[["ind"]]$outcome1_2,
                     iconunicode = "f140")   
fct_plot_indic_donut(indicator = datalist[["ind"]]$birthCertificate,
                     iconunicode = "f140")   
fct_plot_indic_donut(indicator = datalist[["ind"]]$birthRegistered,
                     iconunicode = "f140")   
}
