% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome1_3.R
\name{outcome1_3}
\alias{outcome1_3}
\title{outcome1_3}
\usage{
outcome1_3(datalist)
}
\arguments{
\item{datalist}{A list with all hierarchical data frame for a survey data set.
format is expected to match the Excel export synchronized from kobo to RILD
and loaded with  kobocruncher::kobo_data()}
}
\value{
new calculated variable - appended or not...
}
\description{
\strong{Proportion of Persons of Concern with legally recognized identity documents or credentials}
}
\details{
This indicator measures the proportion of persons of concern who possess legally
recognized and valid identity documents or credentials.

Establishing one's identity and possessing legally recognized and valid identity
documents or credentials ensures the legal protection of persons of concern,
protection from refoulement, registration of life events to prevent statelessness,
as well as access to services.

\strong{Definitions}

\if{html}{\out{<div class="sourceCode">}}\preformatted{Identity document or credential is any document or credential which may be 
used as proof of identity, which may also include reference to the
 individuals' legal status and associated rights vis-à-vis the host 
 State and/or UNHCR.
}\if{html}{\out{</div>}}\tabular{c}{
   Standard Questions \cr
   REG01 - REG02 - REG03 / REG05 - REG06 \cr
}


\strong{Numerator}: Total population with valid identity documents or credentials

\strong{Denominator}: Total population

\strong{Formula}: ( \emph{REG01=1 \| REG02=1 \| REG03=1 \| REG05=1 \| REG06=1) / Total population}

This indicator comes from the individual dataset -
Calculate valid identity documents for under 5 with REG05 and REG06 variables
ind$REG05a - passport
ind$REG05b - civil/government issued ID
ind$REG05c - residency permit
ind$REG05d - statelessness documentation
ind$REG05e - household card of address/family book
ind$REG05f - social security card
ind$REG06 - any other document establishes identity
add birth certificate as additional document from REG03
}
\examples{
## data, cf example  fct_re_map()
datalist <- kobocruncher::kobo_data( system.file("dummy_RMS_CAPI_v2_mapped.xlsx", 
                                                 package = "IndicatorCalc"))
## Apply indicator function on datalist
datalist <- outcome1_3(datalist)

## Visualise value
fct_plot_indic_donut(indicator = datalist[["ind"]]$outcome1_3,
                     iconunicode = "f140")    

fct_plot_indic_donut(indicator = datalist[["ind"]]$document_above5,
                     iconunicode = "f140")    

fct_plot_indic_donut(indicator = datalist[["ind"]]$document_under5,
                     iconunicode = "f140")    
}
