% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome4_1.R
\name{outcome4_1}
\alias{outcome4_1}
\title{outcome4_1}
\usage{
outcome4_1(datalist)
}
\arguments{
\item{datalist}{A list with all hierarchical data frame for a survey data set.
format is expected to match the Excel export synchronized from kobo to RILD
and loaded with  kobocruncher::kobo_data()}
}
\value{
new calculated variable - appended or not...
}
\description{
\strong{Proportion of Persons of Concern who know where to access available GBV services}
}
\details{
The indicator is defined as the proportion of persons of concern who know where to access at least one of the four available gender-based violence services in the aftermath of a GBV incident.

This is linked to the Inter-Agency Minimum Standards for Gender-Based Violence in Emergencies Programming.

The response options will appear as below when the standardized module is used.

\emph{Definitions:}
\enumerate{
\item **GBV01a (**Health): "Medical treatment and health care to address the immediate and long-term physical and mental health effects of GBV. This can include initial examination and treatment, follow-up medical care, mental health care, and health-related legal services, such as preparation of documentation and provision of evidence during judicial and related processes."
\item \strong{GBV01b} (Psycho-social/ case management): "Psychosocial care and support to assist with healing and recovery from emotional, psychological and social effects. This includes crisis care as well as longer-term emotional and practical support for the survivor and her/his family, information and advocacy, case management, and educating family members so that they can support the survivor's healing and recovery.
\item \strong{GBV01c} (Safety/security): "Options for safety and protection for survivors and their families who are at risk of further violence and who wish to be protected. This can include safe shelters, police or community security, relocation, or in the case of children, alternative care arrangements".
\item \strong{GBV01d} (Legal assistance): "Legal actors will clearly and honestly inform the victim/survivor of the #' procedures, limitations, pros, and cons of all existing legal options".
}\tabular{c}{
   Standard Questions \cr
   GBV01 \cr
}


\strong{Numerator}: Total population who indicates knowing where to access available GBV services

\strong{Denominator}: Total population

\strong{Formula}: ( \emph{GBV01a=1 \| GBV01b=1 \| GBV01c=1 \| GBV01d=1) / Total population}

This indicator comes from main dataset based on the respondent randomly selected for individual level
}
\examples{
## data, cf example  fct_re_map()
datalist <- kobocruncher::kobo_data( system.file("dummy_RMS_CAPI_v2_mapped.xlsx", 
                                                 package = "IndicatorCalc"))
## Apply indicator function on datalist
datalist <- outcome4_1(datalist )

## Visualise value
fct_plot_indic_donut(indicator = datalist[["main"]]$outcome1_4,
                     iconunicode = "f140")   
}
