% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome4_2.R
\name{outcome4_2}
\alias{outcome4_2}
\title{outcome4_1}
\usage{
outcome4_2(datalist)
}
\arguments{
\item{datalist}{A list with all hierarchical data frame for a survey data set.
format is expected to match the Excel export synchronized from kobo to RILD
and loaded with  kobocruncher::kobo_data()}
}
\value{
new calculated variable - appended or not...
}
\description{
\strong{Proportion of Persons of Concern who do not accept violence against women}
}
\details{
This indicator measures the proportion of persons of concern who disagree that a husband is justified in hitting or beating his wife for a list of specific reasons.
This serves as a proxy indicator for those who do not accept violence against women.

This module is taken from \href{https://dhsprogram.com/pubs/pdf/DHSQMP/DHS6_Module_Domestic_Violence_6Aug2014_DHSQMP.pdf}{Domestic Violence Module} from DHS and similar module used by MICS.
Due to its sensitivity, this module is only recommended for face-to-face surveys.

\emph{Data Collection Safeguards}

Enumerators (and data collection teams) must be trained on GBV safe disclosures and referrals prior to undertaking household surveys (or at least get a briefing on the GBV referral pathway in their area).
To ensure quality feedback and limit risks of harm, it is advisable to administer the survey to members of the household separately in confidential spaces (consider using interview rooms at community centres in settings where layout of communities' homes do not allow for confidential discussions).

It is also recommended to ensure same-sex enumerators to minimize risks of SEA particularly for female respondents.\tabular{c}{
   Standard Questions \cr
   VAW01 \cr
}


\strong{Numerator}: Population 18 and above who disagree that a husband is justified in hitting or beating his wife for all the following reasons:
\enumerate{
\item Going out without telling him.
\item Neglecting the children.
\item Arguing with him.
\item Refusing to have sexual intercourse with him.
\item Burning food.
}

\strong{Denominator}: Total population 18 and above

\strong{Formula}: ( \emph{GBV01a=1 \| GBV01b=1 \| GBV01c=1 \| GBV01d=1) / Total population}

This indicator comes from main dataset based on the respondent randomly selected for individual level

If randomly selected adult who believes that a  husband is justified in beating his wife in various circumstances

If yes selected for any of the circumstances
Prefer not to respond will be put into missing
}
\examples{
## data, cf example  fct_re_map()
datalist <- kobocruncher::kobo_data( system.file("dummy_RMS_CAPI_v2_mapped.xlsx", 
                                                 package = "IndicatorCalc"))
## Apply indicator function on datalist
datalist <- outcome4_2(datalist)

## Visualise value
fct_plot_indic_donut(indicator = datalist[["main"]]$outcome4_2,
                     iconunicode = "f140")   
}
