% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome5_2.R
\name{outcome5_2}
\alias{outcome5_2}
\title{outcome5_2}
\usage{
outcome5_2(datalist)
}
\arguments{
\item{datalist}{A list with all hierarchical data frame for a survey data set.
format is expected to match the Excel export synchronized from kobo to RILD
and loaded with  kobocruncher::kobo_data()}
}
\value{
new calculated variable - appended or not...
}
\description{
\strong{Proportion of children who participate in community-based child protection programmes}
}
\details{
The indicator is defined as the the proportion of children who participate in community-based child protection programmes.
The module for this question comes from UNHCR Core indicator metadata.

Children' participation in community-based recreational and child protection activities is a key mechanism for the protection of children by providing children with a protected environment in which they can participate in organized activities to play, socialize, learn, and express themselves, promoting their psycho-social well-being and reducing the risk of abuse, violence or exploitation.\tabular{c}{
   Standard Questions \cr
   COMM01-COMM04 \cr
}


\strong{Numerator}: Total number of children between 4-17 years who participate in community-based recreational and child protection programmes

\strong{Denominator}: Total [number of children between 4-17 years

\strong{Formula}: ( \emph{COMM01=1 & COMM02 \>=2 \| & COMM03=1 & COMM04=1) / Total number of children 4-17 years}
This indicator comes from the individual level dataset

Children who participate in community-based programmes at least 2 times, under adult supervision in a physically safe area
}
\examples{
## data, cf example  fct_re_map()
datalist <- kobocruncher::kobo_data( system.file("dummy_RMS_CAPI_v2_mapped.xlsx", 
                                                 package = "IndicatorCalc"))
## Apply indicator function on datalist
datalist <- outcome5_2(datalist  )

## Visualise value
fct_plot_indic_donut(indicator = datalist[["ind"]]$outcome5_2,
                     iconunicode = "f140")   
}
