% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome8_2.R
\name{outcome8_2}
\alias{outcome8_2}
\title{outcome8_2}
\usage{
outcome8_2(datalist)
}
\arguments{
\item{datalist}{A list with all hierarchical data frame for a survey data set.
format is expected to match the Excel export synchronized from kobo to RILD
and loaded with  kobocruncher::kobo_data()}
}
\value{
new calculated variable - appended or not...
}
\description{
\strong{Proportion of Persons of Concern with primary reliance on clean (cooking) fuels and technology}
}
\details{
This indicator measures the proportion of population who use clean fuel and technology as primary source of cooking.
Any type of cooking fuel and technology meeting the WHO set standards on indoor air emissions, as their main cooking solution.
This is also linked to SGD indicator \href{https://sdg-tracker.org/energy}{7.1.2} .

The standard questions are similar to MICS and the calculation of indicators are based on \href{https://mics.unicef.org/files?job=W1siZiIsIjIwMTcvMDIvMDMvMTYvMjcvMjUvNTk5L1BpY3RvcmlhbHNfV0hPX0hvdXNlaG9sZF9FbmVyZ3lfVXNlX0NhdGFsb2d1ZV9TZXB0ZW1iZXJfMjAxNl8ucGRmIl1d&sha=57b4a452fcc0ac88}{UNICEF MICS} and \href{https://www.who.int/data/gho/indicator-metadata-registry/imr-details/5662}{WHO} guidance on the calculation of the same SGD indicator.\tabular{c}{
   Standard Questions \cr
   COOK01- COOK03 \cr
}


\strong{Numerator}: Percentage of population with primary reliance on clean fuels and technologies for cooking

\strong{Denominator}: Total population

\strong{Formula}: COOK01=1 (there is a cooking device) & (COOK02=1,2,3,4,5) \| (COOK02=6 & COOK03=1,2) / Total Population

This indicator comes from household level dataset
Based on MICS calculation : TC4.1
}
\examples{
## data, cf example  fct_re_map()
datalist <- kobocruncher::kobo_data( system.file("dummy_RMS_CAPI_v2_mapped.xlsx", 
                                                 package = "IndicatorCalc"))
## Apply indicator function on datalist
datalist <- outcome8_2(datalist )

## Visualise value
fct_plot_indic_donut(indicator = datalist[["main"]]$outcome8_2,
                     iconunicode = "f140")   
}
