% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome9_1.R
\name{outcome9_1}
\alias{outcome9_1}
\title{outcome9_1}
\usage{
outcome9_1(datalist)
}
\arguments{
\item{datalist}{A list with all hierarchical data frame for a survey data set.
format is expected to match the Excel export synchronized from kobo to RILD
and loaded with  kobocruncher::kobo_data()}
}
\value{
new calculated variable - appended or not...
}
\description{
\strong{Proportion of Persons of Concern living in habitable and affordable housing}
}
\details{
This indicator measures the proportion of persons of concern living in habitable and affordable housing.
This indicator focuses only on habitability and affordability.
It is limited in its reflection on other key aspects of adequate housing including security of tenure, availability of other basic services and infrastructure, accessibility, location of housing, and cultural appropriateness.
It is also linked to SGD indicator \href{https://unstats.un.org/sdgs/metadata/files/Metadata-11-01-01.pdf}{11.1.1}.

The right to access adequate housing is protected by international law.
The concept of "adequacy" means that housing is more than four walls and a roof as indicated in \href{https://spherestandards.org/wp-content/uploads/Sphere-Handbook-2018-EN.pdf}{The Sphere Handbook}.
Habitable housing primarily refers to the fact that the housing should provide protection from cold, damp, heat, rain, wind, and other threats to health, structural hazards, and disease vectors and it should not be overcrowded.
As shelter/housing is primarily a contextual element, there may be discrepancies from country to country on how this data is measured.Habitable shelter is measured based on having improved material for the dwelling as indicated in \href{https://dhsprogram.com/pubs/pdf/AS61/AS61.pdf}{DHS} publication on housing conditions which is also used by \href{https://mics.unicef.org/tools}{MICS6}.
Overcrowding is also used which occurs if there are more than three people per habitable room as defined by \href{https://www.ncbi.nlm.nih.gov/books/NBK535289/table/ch3.tab2/}{UN-Habitat}.

Affordable housing refers to the fact that the cost of housing and its related expenditures on maintenance and household items should be at such a level that it should not compromise the attainment and satisfaction of other basic needs\tabular{c}{
   Standard Questions \cr
   DWE01-DWE05 & DWE08-DWE09 \cr
}


\strong{Numerator}: Total population living in habitable and affordable housing

\strong{Denominator}: Total population

\strong{Formula}: \emph{DWE01} = 1,2 & \emph{DWE02} = 3,4,5,6,7,8,9 & \emph{DWE03} = 8,9,10,11,12,13 & \emph{DWE04} = 10,11,12,13,14,15 & crowding (\emph{HH01}/\emph{DWE05}) \<= 3 & (DWE08=1 & DWE09=1,2) \| DWE08=0

This indicator is calculated from the main dataset

classify as habitable when improved/adequate shelter
}
\examples{
## data, cf example  fct_re_map()
datalist <- kobocruncher::kobo_data( system.file("dummy_RMS_CAPI_v2_mapped.xlsx", 
                                                 package = "IndicatorCalc"))
## Apply indicator function on datalist
datalist <- outcome9_1(datalist)

## Visualise value
fct_plot_indic_donut(indicator = datalist[["main"]]$outcome9_1,
                     iconunicode = "f140")   
}
