% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot_indic_donut.R
\name{fct_plot_indic_donut}
\alias{fct_plot_indic_donut}
\title{fct_plot_indic_donut}
\usage{
fct_plot_indic_donut(
  indicator,
  subtitle_chart = NULL,
  caption_chart = NULL,
  ordered_threhold = NULL,
  iconunicode = "f496"
)
}
\arguments{
\item{indicator}{vector}

\item{subtitle_chart}{}

\item{caption_chart}{}

\item{ordered_threhold}{vector with the different threshold
(green, orange, red)}

\item{iconunicode}{unicode value for fontawesome ---
see https://fontawesome.com/search?o=r&m=free}
}
\value{
ggplot2 object
}
\description{
Wrapper for a summary chart for indicator - allows to add hint, source,  icon
and interpretation threshold if available
}
\examples{
test <- data.frame(
  shelter = rbinom(20, 1, 0.5)) |> 
  dplyr::mutate( shelter = 
  labelled::labelled( shelter,
                      labels = c( "Yes" = 1, "No" = 0),
                      label = "Access to adequate shelter")) 
  
fct_plot_indic_donut(indicator = test$shelter,
                     subtitle_chart = NULL,
                     caption_chart = NULL,
                     ordered_threhold = NULL,
                     iconunicode = "e54f") 
}
