% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_var_mapping.R
\name{fct_var_mapping}
\alias{fct_var_mapping}
\title{fct_var_mapping}
\usage{
fct_var_mapping(xlsformpath)
}
\arguments{
\item{xlsformpath}{path to the xlsform}

\item{mappingfile_out}{path to file to out the mapping}
}
\value{
a list with the best mapping...
}
\description{
Create a diff to identify the mapping between expected variables / modalities and
what is has been configured in a specific xlsform
}
\examples{

form <- system.file("RMSCAPI.xlsx", package = "IndicatorCalc")
#fct_var_mapping(xlsformpath = form)
}
