% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_2_2.R
\name{impact_2_2}
\alias{impact_2_2}
\title{impact_2_2}
\usage{
impact_2_2(
  datalist,
  mapper = list(hierarchy = "main", variablemap = data.frame(label =
    c("Access to shelter", "Access to electricity", "Access to drinking water",
    "Access to  healthcare"), variable = c("shelter", "electricity", "drinkingwater",
    "healthcare"), mappattern = c("shelter", "electricity", "drinkingwater",
    "healthcare")), modalitymap = data.frame(variable = c("shelter", "shelter",
    "electricity", "electricity", "drinkingwater", "drinkingwater", "healthcare",
    "healthcare"), label = c("Yes", "No", "Yes", "No", "Yes", 
     "No", "Yes", "No"),
    standard = c("1", "0", "1", "0", "1", "0", "1", "0"), map = c("1", "0", "1", "0",
    "1", "0", "1", "0")))
)
}
\arguments{
\item{datalist}{A list with all hierarchical data frame for a survey data set.
format is expected to match the Excel export synchronized from kobo to RILD
and loaded with  kobocruncher::kobo_data()}

\item{mapper}{a list providing the mapping of the variables used for the
calculation - this mapper is potentially to be adjusted in relation with deviation
between the the standard XlsForm and the contextualized dataset}
}
\value{
new calculated variable - appended or not...
}
\description{
\strong{Proportion of Persons of Concern residing in physically safe and secure
settlements with access to basic facilities}
}
\details{
Once electricity, healthcare, drinking water and adequate shelter are
calculated, check the values for each variable before,
calculating 2.2 impact indicator.

Once all variables are correctly calculated, we can compute the final
variable for impact 2.2 indicator.
\strong{Numerator}: Population residing in physically safe and secure settlements with access to basic facilities
\strong{Denominator}: Total population
\strong{Formula}: \emph{shelter}=1 & \emph{electricity}=1 & \emph{healthcare}=1 & \emph{drinkingwater}=1

Impact 2.2 is "1" if all services above are accessible
}
\examples{
datalist <- kobocruncher::kobo_data( system.file("test.xlsx",
                                                 package = "IndicatorCalc"))

#Healthcare
mapper <-  list(
            hierarchy = "main",
            variablemap = data.frame(
              label = c(
  "In general, when anyone in your household is sick, 
   where do they go to seek care?",
  "How long does it take to go there when you use the mode of transport 
  that you mentioned above?"),
              variable = c("HEA01", 
                           "HEA03"),
              mappattern = c("HEA01", 
                           "HEA03") ),
            modalitymap = data.frame(
              variable = c( "HEA01", "HEA01" ),
              label = c(  "Other, specify", "Don't know"),
              standard = c("96", "98" ),
              map = c("96", "98" )))

datalist <- inter_healthcare(datalist, mapper )

## Electricity
mapper <- list(
            hierarchy = "main",
            variablemap = data.frame(
              label = c("Does this household use anything for lighting?",
       "What source of electricity is used most of the time in this household?"),
              variable = c("LIGHT01", 
                           "LIGHT03"),
              mappattern = c("LIGHT01", 
                           "LIGHT03") ),
            modalitymap = data.frame(
              variable = c( "LIGHT01", 
                            "LIGHT03", "LIGHT03", "LIGHT03"),
              label = c( "yes",
                        "No electricity in household", "Other, specify", "Don't know"),
              standard = c( "1",
                           "1", "96", "98"),
              map = c("yes",
                      "1", "96", "98"))) 

datalist <- inter_electricity( datalist =datalist, mapper = mapper  )

## Drinking Water
## in the contextualised form - DWA03a has been skipped and all results are in min... 
## only manual transformation can adjust this before we use the mapper..

datalist[["main"]]$DWA03a  <- "1" 

datalist[["main"]]$DWA03b <- 
  datalist[["main"]]$VulnerabilityScoring.BasicNeeds.DWA03

# now the mapper
mapper <-  list(
            hierarchy = "main",
            variablemap = data.frame(
              label = c(
               "What is the main source of drinking water for this household?",
               "Where is this source located?",
               "Unit used to measure time to access",
               "How long does it take to go there, get water, and come back,
                including waiting time?"),
              variable = c("DWA01", 
                           "DWA02", 
                           "DWA03a", 
                           "DWA03b" ),
              mappattern = c("DWA01", 
                           "DWA02", 
                           "DWA03a", 
                           "DWA03b" ) ),
            modalitymap = data.frame(
             variable = c("DWA01",  "DWA01", "DWA01","DWA01",  "DWA01", 
                   "DWA02","DWA02", "DWA02",
                   "DWA03a","DWA03a"),
             label = c( 
                ##DWA01
                "Unprotected Dug Well", 
                "Unprotected Spring",
                "Surface Water (River, Stream, Pond, Dam, Canal)",
                "Other, specify",
                "Don't know",
                ##DWA02
                "In Own Dwelling", 
                "In Own Yard/Plot",
                "Elsewhere",
                ## DWA03a
                "Minutes", 
                "Hours"    ),
             standard = c( "7", "9", "13", "96", "98",
                           "1", "2", "3",
                           "1", "2"),
             map = c( "7", "9", "13", "96", "98",
                       "1", "2", "3",
                        "1", "2") ) 
         )
  
datalist <- inter_drinkingwater(datalist, mapper )

##Shelter
mapper <- list(
            hierarchy = "main",
            variablemap = data.frame(
              label = c(
                "What type of dwelling does the household live in?",
"Main material of the dwelling floor",
"Main material of the roof",
"Main material of the exterior walls",
"How many separate rooms do the members of your household occupy?",
"What is the total number of persons in this household?"),
              variable = c("DWE01","DWE02","DWE03","DWE04","DWE05", 
                           "HH01"),
              mappattern = c("DWE01","DWE02","DWE03","DWE04","DWE05", 
                           "progres_groupsize") ),
            modalitymap = data.frame(
              variable = c( "DWE01","DWE01",
                            "DWE02","DWE02","DWE02",
                            "DWE03","DWE03","DWE03","DWE03","DWE03","DWE03",
                            "DWE04","DWE04","DWE04","DWE04","DWE04","DWE04"),
              label = c(  "Apartment", "House", # DWE01
                         "Earth/sand", "Dung", "Other (Specify)", #DWE02
                         
                         "Metal/tin", "Wood", "Calamine/Cement fibre", 
                         "Ceramic tiles", "Cement", "Roofing shingles",#DWE03
                         
                         "Cement", "Stone with lime/ cement", "Bricks", 
                         "Cement blocks", "Covered adobe", "Wood planks/shingles" # DWE04
                         ),
              standard = c( "1","2",
                           "1", "2","96",
                           "8","9","10","11","12","13",
                           "10","11","12","13","14","15"),
              map = c("1","2",
                           "1", "2","96",
                           "8","9","10","11","12","13",
                           "10","11","12","13","14","15"))) 
## Calculate
datalist <-  inter_shelter(datalist, mapper)

## and now impact

mapper <-   list(
            hierarchy = "main",
            variablemap = data.frame(
              label = c("Access to shelter", 
                        "Access to electricity", 
                         "Access to drinking water", 
                         "Access to  healthcare"),
              variable = c("shelter", 
                           "electricity", 
                           "drinkingwater", 
                           "healthcare"),
              mappattern = c("shelter", 
                           "electricity", 
                           "drinkingwater", 
                           "healthcare") ),
            modalitymap = data.frame(
              variable = c( "shelter", "shelter",
                           "electricity",  "electricity", 
                           "drinkingwater", "drinkingwater", 
                           "healthcare","healthcare"),
              label = c( "Yes","No",
                       "Yes","No",
                       "Yes","No",
                       "Yes","No"),
              standard = c( "1","0",
                           "1","0",
                           "1","0",
                           "1","0"),
              map = c("1","0",
                           "1","0",
                           "1","0",
                           "1","0")))
 
datalist <-  impact_2_2(datalist, mapper)

fct_plot_indic_donut(indicator = datalist[["main"]]$impact2_2,
                     iconunicode = "f140") 
}
